telegram_chatbot/requirements.txt
aiogram
openai
python-dotenv
